WIP: Reimplement the 128Layers test from containerd
This test is a little hacky, I'm committing it to exercise the base
layer creation, and a bunch of other WCOW Layer code, in order to nail
down issues I'm seeing with the containerd integration.

It's also a good place to see where hcsshim still has some places where
it could implement an API to make its users' lives easier, or at least
less dependent on WCOW layer implementation details.

TODO: I left out verifying that the files and directories exist with the
correct content, but I probably should, in case that's why the tests
pass here but fail within containerd.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>